// appfront1/src/modules/publicRoutes/pages/publicDashboardRealTime/interfaces.ts
export interface TeamData {
    key: string;
    team: string;
    countForService: number;
    countInAttendance: number;
    averageTimeValue: number;
    waitingAverageTimeValue: number;
    attendanceAverageTimeValueClose: number;
    waitingAverageTimeValueClose: number;
    countClosedAttendance: number;
    waitingAverageTime: string;
    attendanceAverageTime: string;
    attendanceAverageTimeValue: number;
}